# openapi.yml
openapi: 3.0.3
info:
  title: Fineract Credit Scorecard
  version: 0.1.0-SNAPSHOT
  description: An API module for credit risk assessment
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /api/v1/algorithms/predict:
    post:
      operationId: algorithms_predict
      description: Predict credit risk for a loan
      parameters:
        - in: query
          name: classifier
          schema:
            type: string
          description: The algorithm/classifier to use
          required: true
          examples:
            Example1:
              value: RandomForestClassifier
              summary: Example 1
        - in: query
          name: dataset
          schema:
            type: string
          description: The name of the dataset
          examples:
            Example1:
              value: german
              summary: Example 1
        - in: query
          name: status
          schema:
            type: string
          description: The status of the algorithm
          deprecated: true
          examples:
            Example1:
              value: production
              summary: Example 1
        - in: query
          name: version
          schema:
            type: string
            default: 0.0.1
          description: Algorithm version
          required: true
          examples:
            Example1:
              value: 0.0.1
              summary: Example 1
      tags:
        - algorithms
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
              description: Unspecified request body
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: {}
              description: Unspecified request body
          multipart/form-data:
            schema:
              type: object
              additionalProperties: {}
              description: Unspecified request body
      security:
        - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionResponse'
          description: ''
...
components:
  schemas:
...
    PredictionResponse:
      type: object
      properties:
        probability:
          type: number
          format: float
        label:
          type: string
        method:
          type: string
        color:
          type: string
        wilkis_lambda:
          type: number
          format: float
        pillais_trace:
          type: number
          format: float
        hotelling_tawley:
          type: number
          format: float
        roys_reatest_roots:
          type: number
          format: float
        request_id:
          type: integer
      required:
        - color
        - hotelling_tawley
        - label
        - method
        - pillais_trace
        - probability
        - request_id
        - roys_reatest_roots
        - wilkis_lambda
servers:
  - url: http://127.0.0.1:8000
    description: server on localhost